--- a/aptos-move/aptos-gas-schedule/src/gas_schedule/transaction.rs
+++ b/aptos-move/aptos-gas-schedule/src/gas_schedule/transaction.rs
@@ -257,7 +257,7 @@ crate::gas_schedule::macros::define_gas_parameters!(
         [
             keyless_base_cost: InternalGas,
             { RELEASE_V1_12.. => "keyless.base" },
-            414_000_000,
+            138_000_000,
         ]
     ]
 );