## Description

There have been 2 performance improvements to keyless TXN verification times:
1. Caching the deserialized VK in the VM (https://github.com/aptos-labs/aptos-core/pull/12975), which lowered the time from 5ms to 3.10 ms
2. Faster Poseidon hashing (https://github.com/aptos-labs/aptos-core/pull/13050), which lowered the gas from 3.10ms to 1.66ms

Overall, this is a ~3x reduction in gas costs and should be reflected in how we charge for keyless TXNs.

## Type of Change
- [ ] New feature
- [ ] Bug fix
- [ ] Breaking change
- [x] Performance improvement
- [ ] Refactoring
- [ ] Dependency update
- [ ] Documentation update
- [ ] Tests

## Which Components or Systems Does This Change Impact?
- [x] Validator Node
- [ ] Full Node (API, Indexer, etc.)
- [x] Move/Aptos Virtual Machine
- [ ] Aptos Framework
- [ ] Aptos CLI/SDK
- [ ] Developer Infrastructure
- [ ] Other (specify)

## How Has This Been Tested?

It has not. @vgao1996 do we need to bump up gas versions?

## Key Areas to Review

1. Do we need to bump up gas version?
2. Do we need feature gating?

## Checklist
- [x] I have read and followed the [CONTRIBUTING](https://github.com/aptos-labs/aptos-core/blob/main/CONTRIBUTING.md) doc
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I identified and added all stakeholders and component owners affected by this change as reviewers
- [x] I tested both happy and unhappy path of the functionality
- [x] I have made corresponding changes to the documentation

<!-- Thank you for your contribution! -->