@alinush 2024-05-23
> No need to bump the gas feature version if you're just changing the value

Cool, thank you @vgao1996! Can you stamp please? 🙏